requirements.txt
scikit-learn==1.3.0
numpy==1.24.3
Pillow==10.0.0
Flask==2.3.3
